refactor: 페이지 폴더 하위에 api, queries 생성 및 관련 파일 이동
